// 14-Ft_transcendence/backend-nest/src/game/dto/invite.dto.ts
/* ************************************************************************** */
/*                                                                            */
/*                                                        :::      ::::::::   */
/*   invite.dto.ts                                      :+:      :+:    :+:   */
/*                                                    +:+ +:+         +:+     */
/*   By: mmarinel <mmarinel@student.42.fr>          +#+  +:+       +#+        */
/*                                                +#+#+#+#+#+   +#+           */
/*   Created: 2023/11/19 13:40:45 by mmarinel          #+#    #+#             */
/*   Updated: 2023/11/19 13:40:46 by mmarinel         ###   ########.fr       */
/*                                                                            */
/* ************************************************************************** */

export class InviteDto {
	hostID: number
	guestID: number
}